styles for screener tooltips, updated san-ui
