+++ package.json
@@ -18,7 +18,7 @@
   "dependencies": {
     "@apollo/react-testing": "^3.1.3",
     "@santiment-network/chart": "^0.5.0",
-    "@santiment-network/ui": "^0.5.189",
+    "@santiment-network/ui": "^0.5.193",
     "@trainline/react-skeletor": "^1.0.2",
     "apollo-cache-inmemory": "^1.6.3",
     "apollo-client": "^2.6.4",
