+++ src/components/Navbar/ChartLayouts/NavbarChartsDropdown.module.scss
@@ -22,7 +22,7 @@
   color: var(--waterloo);
   display: flex;
   justify-content: space-between;
-  margin: 0 0 8px 8px;
+  margin: 0 0 0 8px;
 }
 
 .list {
@@ -33,7 +33,12 @@
 .scroll {
   margin-top: 14px;
   height: calc(100% - 24px - 14px);
-  overflow-y: auto;
+  overflow-y: hidden;
+  padding-right: 4px;
+
+  &:hover {
+    overflow-y: auto;
+  }
 }
 
 .btn {
