+++ src/components/Navbar/ChartLayouts/NavbarChartsLayouts.module.scss
@@ -3,17 +3,15 @@
 .wrapper {
   box-sizing: border-box;
   flex: 1;
-  overflow-y: auto;
   overflow-x: hidden;
   position: relative;
   margin-top: 14px;
-}
+  overflow-y: hidden;
+  padding-right: 4px;
 
-.list {
-  position: absolute;
-  left: 0;
-  top: 0;
-  width: 100%;
+  &:hover {
+    overflow-y: auto;
+  }
 }
 
 .item {
