+++ src/ducks/Signals/ScreenerSignal/ScreenerSignalDialog.module.scss
@@ -8,7 +8,7 @@
 
   &:hover {
     svg {
-      fill: var(--jungle-green);
+      fill: var(--jungle-green-light-3);
     }
   }
 }
