+++ src/ducks/Watchlists/Widgets/VolumeChart/ProjectsChart.module.scss
@@ -91,3 +91,8 @@
   margin-top: 16px;
   height: 128px;
 }
+
+.tooltipHeader {
+  background: var(--jungle-green-light-2);
+  color: var(--rhino);
+}
